roles/frontend/app/requirements.txt
streamlit
google-cloud-storage
tensorflow==2.3.0